DDCNL-7801: Moved authWithDataRetrieval setup to BaseSpec
